[IMP] product, stock: move ZPL product labels, merge unpriced labels
ZPL product labels can be used in absence of any move data, meaning
there's no reason to keep them separate from PDF product labels in
`product`.

Task ID: 4277792
