+++ addons/product/report/product_label_report.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
+import markupsafe
 
 from collections import defaultdict
 
@@ -52,6 +51,7 @@ def _prepare_data(env, docids, data):
     return {
         'quantity': quantity_by_product,
         'page_numbers': (total - 1) // (layout_wizard.rows * layout_wizard.columns) + 1,
+        'price_included': data.get('price_included'),
         'extra_html': layout_wizard.extra_html,
         'pricelist': layout_wizard.pricelist_id,
     }
@@ -87,3 +87,45 @@ class ReportProductReport_Producttemplatelabel_Dymo(models.AbstractModel):
 
     def _get_report_values(self, docids, data):
         return _prepare_data(self.env, docids, data)
+
+
+class ReportProductReport_Producttemplatelabel_Zpl(models.AbstractModel):
+    _name = 'report.product.report_producttemplatelabel_zpl'
+    _description = 'Product Label Report ZPL'
+
+    def _get_report_values(self, docids, data):
+        if data.get('active_model') == 'product.template':
+            Product = self.env['product.template']
+        elif data.get('active_model') == 'product.product':
+            Product = self.env['product.product']
+        else:
+            raise UserError(self.env._('Product model not defined, Please contact your administrator.'))
+
+        quantity_by_product = defaultdict(list)
+        for p, q in data.get('quantity_by_product').items():
+            product = Product.browse(int(p))
+            default_code_markup = markupsafe.Markup(product.default_code) if product.default_code else ''
+            product_info = {
+                'barcode': markupsafe.Markup(product.barcode) if product.barcode else '',
+                'quantity': q,
+                'display_name_markup': markupsafe.Markup(product.display_name),
+                'default_code': (default_code_markup[:15], default_code_markup[15:30])
+            }
+            quantity_by_product[product].append(product_info)
+        if data.get('custom_barcodes'):
+            # we expect custom barcodes to be: {product: [(barcode, qty_of_barcode)]}
+            for product, barcodes_qtys in data.get('custom_barcodes').items():
+                product = Product.browse(int(product))
+                default_code_markup = markupsafe.Markup(product.default_code) if product.default_code else ''
+                for barcode_qty in barcodes_qtys:
+                    quantity_by_product[product].append({
+                        'barcode': markupsafe.Markup(barcode_qty[0]),
+                        'quantity': barcode_qty[1],
+                        'display_name_markup': markupsafe.Markup(product.display_name),
+                        'default_code': (default_code_markup[:15], default_code_markup[15:30])
+                    })
+        data['quantity'] = quantity_by_product
+        layout_wizard = self.env['product.label.layout'].browse(data.get('layout_wizard'))
+        data['pricelist'] = layout_wizard.pricelist_id
+
+        return data
