+++ addons/product/report/product_reports.xml
@@ -93,5 +93,15 @@
             <field name="print_report_name">'Products Labels - %s' % (object.name)</field>
             <field name="binding_type">report</field>
         </record>
+
+        <record id="report_product_template_label_zpl" model="ir.actions.report">
+            <field name="name">Product Label (ZPL)</field>
+            <field name="model">product.product</field>
+            <field name="report_type">qweb-text</field>
+            <field name="report_name">product.report_producttemplatelabel_zpl</field>
+            <field name="report_file">product.report_producttemplatelabel_zpl</field>
+            <field name="binding_model_id" eval="False"/>
+            <field name="binding_type">report</field>
+        </record>
     </data>
 </odoo>
