+++ addons/product/report/product_template_templates.xml
@@ -34,4 +34,8 @@
     </t>
 </template>
 
+<template id="report_producttemplatelabel_zpl">
+    <t t-call="product.report_productlabel_zpl"/>
+</template>
+
 </odoo>
