+++ addons/product/wizard/product_label_layout.py
@@ -1,21 +1,41 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 
 from odoo import _, api, fields, models
 from odoo.exceptions import UserError
+from odoo.tools import BinaryBytes, file_open
+
+# Format table expressed as width x height in inch.
+ZPL_FORMAT_SIZE = {
+    'normal': (2.25, 1.25),
+    'small': (1.25, 1.00),
+    'alternative': (2.00, 1.00),
+    'jewelry': (2.20, 0.50),
+}
 
 
 class ProductLabelLayout(models.TransientModel):
     _name = 'product.label.layout'
     _description = 'Choose the sheet layout to print the labels'
 
+    @api.model
+    def _get_zpl_label_placeholder(self):
+        with file_open('product/static/img/zpl_label_placeholder.png', 'rb') as f:
+            return BinaryBytes(f.read())
+
     print_format = fields.Selection([
         ('dymo', 'Dymo'),
         ('2x7', '2 x 7'),
         ('4x7', '4 x 7'),
         ('4x12', '4 x 12'),
+        ('zpl', 'ZPL Labels'),
     ], string="Format", default='2x7', required=True)
+    zpl_template = fields.Selection([
+        ('normal', 'Normal (2.25" x 1.25")'),
+        ('small', 'Small (1.25" x 1.00")'),
+        ('alternative', 'Alternative (2.00" x 1.00")'),
+        ('jewelry', 'Jewelry (2.20" x 0.50")'),
+    ], string="ZPL Template", default='normal', required=True)
+    zpl_preview = fields.Image('ZPL Preview', readonly=True, default=_get_zpl_label_placeholder)
     with_price = fields.Boolean('Print With Price', default=True)
     custom_quantity = fields.Integer('Copies', default=1, required=True)
     product_ids = fields.Many2many('product.product')
@@ -57,6 +77,7 @@ class ProductLabelLayout(models.TransientModel):
             'quantity_by_product': {p: self.custom_quantity for p in products},
             'layout_wizard': self.id,
             'price_included': self.with_price,
+            'zpl_template': self.zpl_template,
         }
         return xml_id, data
