+++ addons/stock/wizard/product_label_layout.py
@@ -1,47 +1,19 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from collections import defaultdict
-from odoo import api, fields, models
-from odoo.tools import BinaryBytes, file_open
-
-# Format table expressed as width x height in inch.
-ZPL_FORMAT_SIZE = {
-    'normal': (2.25, 1.25),
-    'small': (1.25, 1.00),
-    'alternative': (2.00, 1.00),
-    'jewelry': (2.20, 0.50),
-}
+from odoo import fields, models
 
 
 class ProductLabelLayout(models.TransientModel):
     _inherit = 'product.label.layout'
 
-    @api.model
-    def _get_zpl_label_placeholder(self):
-        with file_open('stock/static/img/zpl_label_placeholder.png', 'rb') as f:
-            return BinaryBytes(f.read())
-
     move_ids = fields.Many2many('stock.move')
     move_quantity = fields.Selection([
         ('move', 'Operation Quantities'),
         ('custom', 'Custom')], string="Quantity to print", required=True, default='custom')
-    print_format = fields.Selection(selection_add=[
-        ('zpl', 'ZPL Labels'),
-    ], ondelete={'zpl': 'set default'})
-    zpl_template = fields.Selection([
-        ('normal', 'Normal (2.25" x 1.25")'),
-        ('small', 'Small (1.25" x 1.00")'),
-        ('alternative', 'Alternative (2.00" x 1.00")'),
-        ('jewelry', 'Jewelry (2.20" x 0.50")'),
-    ], string="ZPL Template", default='normal', required=True)
-    zpl_preview = fields.Image('ZPL Preview', readonly=True, default=_get_zpl_label_placeholder)
 
     def _prepare_report_data(self):
         xml_id, data = super()._prepare_report_data()
-
-        if 'zpl' in self.print_format:
-            xml_id = 'stock.label_product_product'
-            data['zpl_template'] = self.zpl_template
 
         quantities = defaultdict(int)
         uom_unit = self.env.ref('uom.product_uom_unit', raise_if_not_found=False)
